feat!: use f64 for number settings (#3)
* feat!: use f64 for number settings

* fix: use f64 in deserialization

+++ generate-mods/src/model.rs
@@ -547,7 +547,7 @@ impl SettingType {
     pub fn write(self, writer: &mut Writer) -> GenResult {
         match self {
             SettingType::Bool => writer.write("bool"),
-            SettingType::Number => writer.write("f32"),
+            SettingType::Number => writer.write("f64"),
             SettingType::String => writer.write("String"),
         }
     }
